requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pandas==2.1.4
numpy==1.26.4
scikit-learn==1.3.2
sentence-transformers==2.2.2
nltk==3.8.1
PyPDF2==3.0.1
python-docx==0.8.11
scipy==1.12.0
torch>=2.0.0
